Move `common.enum_proto` to `common.proto`
We will start to expose conversion functions, for which we will always
use a sub-module called `proto` for these conversion functions, so we
we do the same with `enum_proto`.

The old version is kept as deprecated.
