+++ src/frequenz/client/common/microgrid/electrical_components/__init__.py
@@ -89,7 +89,7 @@ class ElectricalComponentCategory(enum.Enum):
     """A heating, ventilation, and air conditioning (HVAC) system."""
 
     @classmethod
-    @deprecated("Use `frequenz.client.common.enum_proto.enum_from_proto` instead.")
+    @deprecated("Use `frequenz.client.common.proto.enum_from_proto` instead.")
     def from_proto(
         cls, component_category: PBElectricalComponentCategory.ValueType
     ) -> ElectricalComponentCategory:
@@ -217,7 +217,7 @@ class ElectricalComponentStateCode(enum.Enum):
     """The precharger circuit is closed, allowing full current to flow to the main circuit."""
 
     @classmethod
-    @deprecated("Use `frequenz.client.common.enum_proto.enum_from_proto` instead.")
+    @deprecated("Use `frequenz.client.common.proto.enum_from_proto` instead.")
     def from_proto(
         cls, component_state: PBElectricalComponentStateCode.ValueType
     ) -> ElectricalComponentStateCode:
@@ -432,7 +432,7 @@ class ElectricalComponentDiagnosticCode(enum.Enum):
     times."""
 
     @classmethod
-    @deprecated("Use `frequenz.client.common.enum_proto.enum_from_proto` instead.")
+    @deprecated("Use `frequenz.client.common.proto.enum_from_proto` instead.")
     def from_proto(
         cls, component_error_code: PBElectricalComponentDiagnosticCode.ValueType
     ) -> ElectricalComponentDiagnosticCode:
