+++ src/frequenz/client/common/proto/__init__.py
@@ -1,0 +1,10 @@
+# License: MIT
+# Copyright © 2025 Frequenz Energy-as-a-Service GmbH
+
+"""General utilities for converting common types to/from protobuf types."""
+
+from ._enum import enum_from_proto
+
+__all__ = [
+    "enum_from_proto",
+]
